issue #152 refactor: CustomDialog 분리
